correcciones de estilo
